Praise Mullen Matthias!
After implementing position-based-fluids on gpu, can now simulate 250,000+ particles without any noticeable frame rate drops.

+++ Graphics/Shader.cpp
@@ -54,7 +54,7 @@ unsigned int Shader::loadShaderFromFile(GLenum type, const char* fileName) {
 				includeFile.open(name, std::ios::in | std::ios::binary);
 				
 				includeFile.seekg(0, includeFile.end);
-				int fileLength = includeFile.tellg();
+				int fileLength = (int)includeFile.tellg();
 				includeFile.seekg(0, includeFile.beg);
 
 				char* buffer = new char[fileLength + 1];
